refactor: remove redundant src/ shims and dead duplicates
Eliminate 5 one-line re-export shims and 2 fully-duplicated/orphaned
files left behind by an earlier restructure. Net delta: 8 files removed,
1 directory removed, 13 import lines rewritten across 11 files. No
behavior change; boundary rules unchanged.

Removed:
- src/app/services/docs-build-service.ts (DocsBuildService had zero
  importers anywhere)
- src/server-config.ts + tests/server-config.test.ts (DEFAULT_SERVER_PORT,
  MAX_TCP_PORT, parsePort already live verbatim in
  src/adapters/infra/config/env.ts, the actually-wired version)
- src/adapters/mcp/tool-contracts.ts (1-line shim; canonical 14KB Zod
  module is at src/mcp/tool-contracts.ts)
- src/docs/generate.ts (1-line shim -> src/adapters/docs/generate.ts)
- src/docs/projection.ts (1-line shim -> src/core/documents.ts)
- src/logging/runtime-logger.ts (1-line shim -> src/adapters/infra/logging)
- src/observability/runtime-observability.ts (1-line shim -> src/adapters/infra/observability)
- src/observability/ directory (now empty)

PR template updated to point at src/mcp/tool-contracts.ts (the actual
canonical location, not the deleted shim path).

Verified locally: bun run check, check:boundaries, lint, test (170 pass;
1 pre-existing rspress-build env failure unchanged), build (cli.js +
server.js + stdio.js produced), docs:generate (826 files via the
rewritten import path).
