--- src/adapters/mcp/tool-contracts.ts
@@ -1,1 +1,0 @@
-export * from '../../mcp/tool-contracts.js';
