--- src/docs/generate.ts
@@ -1,1 +1,0 @@
-export * from '../adapters/docs/generate.js';
