--- src/docs/projection.ts
@@ -1,1 +1,0 @@
-export * from '../core/documents.js';
