--- src/logging/runtime-logger.ts
@@ -1,1 +1,0 @@
-export * from '../adapters/infra/logging/runtime-logger.js';
