--- src/observability/runtime-observability.ts
@@ -1,1 +1,0 @@
-export * from '../adapters/infra/observability/runtime-observability.js';
